tests: use embedded postgres
