+++ Cargo.toml
@@ -15,7 +15,6 @@ async-trait = "0.1.83"
 clap = "4.5.20"
 toml = "0.8.19"
 sqlx = { version = "0.7.4", default-features = false }
-serial_test = "3.2.0"
 
 # bitcoin
 bitcoin = "0.32.5"
