+++ core/src/verifier.rs
@@ -806,7 +806,7 @@ impl Verifier {
 
 // #[tokio::test]
 // async fn verifier_new_public_key_check() {
-//     let mut config = create_test_config_with_thread_name!(None);
+//     let mut config = create_test_config_with_thread_name(None);
 //     let rpc = ExtendedRpc::new(
 //         config.bitcoin_rpc_url.clone(),
 //         config.bitcoin_rpc_user.clone(),
@@ -823,9 +823,8 @@ impl Verifier {
 // }
 
 // #[tokio::test]
-// #[serial_test::serial]
 // async fn new_deposit_nonce_checks() {
-//     let config = create_test_config_with_thread_name!(None);
+//     let config = create_test_config_with_thread_name(None);
 //     let rpc = ExtendedRpc::new(
 //         config.bitcoin_rpc_url.clone(),
 //         config.bitcoin_rpc_user.clone(),
