+++ core/tests/deposit.rs
@@ -14,9 +14,8 @@
 
 // #[ignore = "We are switching to gRPC"]
 // #[tokio::test]
-// #[serial_test::serial]
 // async fn deposit_with_retry_checks() {
-// let config = create_test_config_with_thread_name!(None);
+// let config = create_test_config_with_thread_name(None);
 // let rpc = ExtendedRpc::new(
 //     config.bitcoin_rpc_url.clone(),
 //     config.bitcoin_rpc_user.clone(),
