+++ core/tests/taproot.rs
@@ -8,16 +8,14 @@ use clementine_core::actor::Actor;
 use clementine_core::builder::transaction::TxHandler;
 use clementine_core::builder::{self};
 use clementine_core::extended_rpc::ExtendedRpc;
+use clementine_core::testkit::create_test_setup;
 use clementine_core::utils::{handle_taproot_witness_new, SECP};
-use clementine_core::{config::BridgeConfig, database::Database, utils::initialize_logger};
-use std::{env, thread};
 
 mod common;
 
 #[tokio::test]
-#[serial_test::serial]
 async fn create_address_and_transaction_then_sign_transaction() {
-    let config = create_test_config_with_thread_name!(None);
+    let (config, _db) = create_test_setup().await.expect("test setup");
     let rpc = ExtendedRpc::new(
         config.bitcoin_rpc_url,
         config.bitcoin_rpc_user,
